+++ qtypes/examples/__main__.py
@@ -4,6 +4,7 @@
 import sys
 
 
+from .dynamic import *
 from .units import *
 from .one_of_each import *