+++ CHANGELOG.md
@@ -28,6 +28,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Added
 - initial release
 
-[Unreleased]: https://gitlab.com/yaq/qtypes/-/compare/v0.2.0...master
+[Unreleased]: https://gitlab.com/yaq/qtypes/-/compare/v0.2.0...main
 [0.2.0]: https://gitlab.com/yaq/qtypes/-/compare/v0.1.0...v0.2.0
 [0.1.0]: https://gitlab.com/yaq/qtypes/-/tags/v0.1.0